requirements.txt
openai~=0.26.5
python-dotenv~=0.21.1